feat(substrate,component): widen kind id u32 → u64 (ADR-0030 Phase 1)
Mechanical widen of `MailKind` and every FFI/registry/test touchpoint
ahead of Phase 2's switch to hashed derivation. Mirrors the ADR-0029
mailbox-id split: this PR only changes the integer width, ids remain
sequential (`kind_names.len() as u64`).

- `MailKind = u64`, `KIND_NOT_FOUND = u64::MAX`
- `send_mail_p32` / `reply_mail_p32` kind params → i64 ABI;
  `resolve_kind_p32` result → i64
- `Registry::{kind_id, kind_name, kind_descriptor, register_kind*}`
  now work in u64
- `KindId<K>.raw`, `Sink<K>.kind`, `Mail::kind`, `KindTable` entries → u64
- `Component::receive` typed func widens to `(u64, u32, u32, u32) -> u32`
- WAT receive_p32 signatures `(param i32 …)` → `(param i64 …)`; WAT
  send/reply_mail kind args `(i32.const …)` → `(i64.const …)`
- `aether-substrate/main.rs` cached kind ids → u64

+++ crates/aether-component/src/kinds.rs
@@ -195,7 +195,7 @@ struct KindTableInner {
 #[derive(Copy, Clone)]
 struct Entry {
     type_id: Option<TypeId>,
-    raw: u32,
+    raw: u64,
 }
 
 impl KindTable {
@@ -220,7 +220,7 @@ impl KindTable {
     /// Caller must guarantee no concurrent access. Intended to be
     /// called only from `KindList::resolve_all`, which the `export!`
     /// macro's init shim invokes before any reads.
-    pub unsafe fn insert(&self, type_id: TypeId, raw: u32) {
+    pub unsafe fn insert(&self, type_id: TypeId, raw: u64) {
         let inner = unsafe { &mut *self.inner.get() };
         if inner.len >= MAX_KINDS {
             panic!("aether-component: KindTable overflow (>{MAX_KINDS} kinds)");
@@ -234,7 +234,7 @@ impl KindTable {
 
     /// Linear-scan lookup. Returns `Some(raw)` if `type_id` was
     /// inserted, `None` otherwise.
-    pub fn lookup(&self, type_id: TypeId) -> Option<u32> {
+    pub fn lookup(&self, type_id: TypeId) -> Option<u64> {
         let inner = unsafe { &*self.inner.get() };
         for i in 0..inner.len {
             let entry = inner.entries[i];
