+++ crates/aether-substrate/src/control.rs
@@ -767,7 +767,7 @@ mod tests {
     const WAT: &str = r#"
         (module
             (memory (export "memory") 1)
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0))
     "#;
 
@@ -778,7 +778,7 @@ mod tests {
     const WAT_HOOKS: &str = r#"
         (module
             (memory (export "memory") 1)
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0)
             (func (export "on_replace") (result i32)
                 i32.const 200
@@ -797,7 +797,7 @@ mod tests {
     const WAT_TRAPS_ON_DROP: &str = r#"
         (module
             (memory (export "memory") 1)
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0)
             (func (export "on_drop") (result i32)
                 unreachable))
@@ -811,7 +811,7 @@ mod tests {
                 (func $save_state (param i32 i32 i32) (result i32)))
             (memory (export "memory") 1)
             (data (i32.const 300) "\de\ad\be\ef")
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0)
             (func (export "on_replace") (result i32)
                 (drop (call $save_state
@@ -829,7 +829,7 @@ mod tests {
             (import "aether" "save_state_p32"
                 (func $save_state (param i32 i32 i32) (result i32)))
             (memory (export "memory") 1)
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0)
             (func (export "on_replace") (result i32)
                 (drop (call $save_state
@@ -845,7 +845,7 @@ mod tests {
     const WAT_REHYDRATES: &str = r#"
         (module
             (memory (export "memory") 1)
-            (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+            (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                 i32.const 0)
             (func (export "on_rehydrate_p32") (param i32 i32 i32) (result i32)
                 i32.const 396
@@ -973,7 +973,7 @@ mod tests {
             r#"(module
                 (@custom "aether.kinds" "{escaped}")
                 (memory (export "memory") 1)
-                (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+                (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                     i32.const 0))"#
         );
         let wasm = wat::parse_str(wat).unwrap();
@@ -1032,7 +1032,7 @@ mod tests {
             r#"(module
                 (@custom "aether.kinds" "{escaped}")
                 (memory (export "memory") 1)
-                (func (export "receive_p32") (param i32 i32 i32 i32) (result i32)
+                (func (export "receive_p32") (param i64 i32 i32 i32) (result i32)
                     i32.const 0))"#
         );
         let wasm = wat::parse_str(wat).unwrap();
@@ -1910,14 +1910,14 @@ mod tests {
     const WAT_FORWARDS_TO_SINK: &str = r#"
         (module
             (import "aether" "send_mail_p32"
-                (func $send_mail (param i64 i32 i32 i32 i32) (result i32)))
+                (func $send_mail (param i64 i64 i32 i32 i32) (result i32)))
             (memory (export "memory") 1)
             (func (export "receive_p32")
-                (param $kind i32) (param $ptr i32) (param $count i32) (param $sender i32)
+                (param $kind i64) (param $ptr i32) (param $count i32) (param $sender i32)
                 (result i32)
                 (drop (call $send_mail
                     (i64.load (local.get $ptr))
-                    (i32.const 0)
+                    (i64.const 0)
                     (i32.const 0)
                     (i32.const 0)
                     (local.get $count)))
